+++ cracker-client/src/markers/description/+components/Image.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { useCurrentLanguage } from "../../+hooks/useCurrentLanguage";
-import { useImage } from "../../+hooks/useImage";
+import { useCurrentLanguage } from "../../../+hooks/useCurrentLanguage";
+import { useImage } from "../../../+hooks/useImage";
 
 export const Image = ({ marker, width, height }) => {
   const { imageFilename } = marker;
